Refactor: convert strategies.py to package grouped by leg count
Split the monolithic strategies.py (1,024 lines) into a strategies/ package
with modules organized by leg count: singles, two_leg, butterflies,
iron_strategies, and calendar. Internal helpers and the Side enum live in
_helpers.py. All 28 public strategy functions are re-exported from
__init__.py so existing imports remain unchanged.
