+++ optopsy/strategies/_helpers.py
@@ -1,0 +1,260 @@
+"""Internal helpers for options strategy processing.
+
+Contains the ``Side`` enum, default parameter dictionaries, and private helper
+functions that assemble leg definitions and call the core strategy engine.
+These are not part of the public API.
+"""
+
+from enum import Enum
+from typing import Any, Dict, List, Tuple
+
+import pandas as pd
+from typing_extensions import Unpack
+
+from ..core import _process_calendar_strategy, _process_strategy
+from ..definitions import (
+    calendar_spread_external_cols,
+    calendar_spread_internal_cols,
+    diagonal_spread_external_cols,
+    diagonal_spread_internal_cols,
+    double_strike_external_cols,
+    double_strike_internal_cols,
+    quadruple_strike_external_cols,
+    quadruple_strike_internal_cols,
+    single_strike_external_cols,
+    single_strike_internal_cols,
+    straddle_internal_cols,
+    triple_strike_external_cols,
+    triple_strike_internal_cols,
+)
+from ..rules import (
+    _rule_butterfly_strikes,
+    _rule_expiration_ordering,
+    _rule_iron_butterfly_strikes,
+    _rule_iron_condor_strikes,
+    _rule_non_overlapping_strike,
+)
+from ..types import StrategyParams
+
+default_kwargs: Dict[str, Any] = {
+    "dte_interval": 7,
+    "max_entry_dte": 90,
+    "exit_dte": 0,
+    "exit_dte_tolerance": 0,
+    "otm_pct_interval": 0.05,
+    "max_otm_pct": 0.5,
+    "min_bid_ask": 0.05,
+    "drop_nan": True,
+    "raw": False,
+    # Greeks filtering (optional)
+    "delta_min": None,
+    "delta_max": None,
+    # Greeks grouping (optional)
+    "delta_interval": None,
+    # Pre-computed signal dates (optional) — use apply_signal() to generate
+    "entry_dates": None,
+    "exit_dates": None,
+    # Slippage settings
+    "slippage": "mid",  # "mid", "spread", or "liquidity"
+    "fill_ratio": 0.5,  # Base fill ratio for liquidity mode (0.0-1.0)
+    "reference_volume": 1000,  # Volume threshold for liquid options
+}
+
+# Calendar strategies share most defaults but don't use delta or max_entry_dte,
+# and override exit_dte with a different default.
+_calendar_only_keys = {"max_entry_dte", "delta_min", "delta_max", "delta_interval"}
+calendar_default_kwargs: Dict[str, Any] = {
+    **{k: v for k, v in default_kwargs.items() if k not in _calendar_only_keys},
+    "front_dte_min": 20,
+    "front_dte_max": 40,
+    "back_dte_min": 50,
+    "back_dte_max": 90,
+    "exit_dte": 7,
+}
+
+
+class Side(Enum):
+    """Enum representing long or short position side with multiplier values."""
+
+    long = 1
+    short = -1
+
+
+def _singles(
+    data: pd.DataFrame, leg_def: List[Tuple], **kwargs: Unpack[StrategyParams]
+) -> pd.DataFrame:
+    """Process single-leg option strategies (calls or puts)."""
+    params = {**default_kwargs, **kwargs}
+    return _process_strategy(
+        data,
+        internal_cols=single_strike_internal_cols,
+        external_cols=single_strike_external_cols,
+        leg_def=leg_def,
+        params=params,
+    )
+
+
+def _straddles(
+    data: pd.DataFrame, leg_def: List[Tuple], **kwargs: Unpack[StrategyParams]
+) -> pd.DataFrame:
+    """Process straddle strategies (call and put at same strike)."""
+    params = {**default_kwargs, **kwargs}
+
+    return _process_strategy(
+        data,
+        internal_cols=straddle_internal_cols,
+        external_cols=single_strike_external_cols,
+        leg_def=leg_def,
+        join_on=[
+            "underlying_symbol",
+            "expiration",
+            "strike",
+            "dte_entry",
+            "dte_range",
+            "otm_pct_range",
+            "underlying_price_entry",
+        ],
+        params=params,
+    )
+
+
+def _strangles(
+    data: pd.DataFrame, leg_def: List[Tuple], **kwargs: Unpack[StrategyParams]
+) -> pd.DataFrame:
+    """Process strangle strategies (call and put at different strikes)."""
+    params = {**default_kwargs, **kwargs}
+    return _process_strategy(
+        data,
+        internal_cols=double_strike_internal_cols,
+        external_cols=double_strike_external_cols,
+        leg_def=leg_def,
+        rules=_rule_non_overlapping_strike,
+        join_on=["underlying_symbol", "expiration", "dte_entry", "dte_range"],
+        params=params,
+    )
+
+
+def _spread(
+    data: pd.DataFrame, leg_def: List[Tuple], **kwargs: Unpack[StrategyParams]
+) -> pd.DataFrame:
+    """Process vertical spread strategies (call or put spreads at different strikes)."""
+    params = {**default_kwargs, **kwargs}
+    return _process_strategy(
+        data,
+        internal_cols=double_strike_internal_cols,
+        external_cols=double_strike_external_cols,
+        leg_def=leg_def,
+        rules=_rule_non_overlapping_strike,
+        join_on=["underlying_symbol", "expiration", "dte_entry", "dte_range"],
+        params=params,
+    )
+
+
+def _butterfly(
+    data: pd.DataFrame, leg_def: List[Tuple], **kwargs: Unpack[StrategyParams]
+) -> pd.DataFrame:
+    """Process butterfly strategies (3 legs at different strikes)."""
+    params = {**default_kwargs, **kwargs}
+    return _process_strategy(
+        data,
+        internal_cols=triple_strike_internal_cols,
+        external_cols=triple_strike_external_cols,
+        leg_def=leg_def,
+        rules=_rule_butterfly_strikes,
+        join_on=["underlying_symbol", "expiration", "dte_entry", "dte_range"],
+        params=params,
+    )
+
+
+def _iron_condor(
+    data: pd.DataFrame, leg_def: List[Tuple], **kwargs: Unpack[StrategyParams]
+) -> pd.DataFrame:
+    """Process iron condor strategies (4 legs at different strikes)."""
+    params = {**default_kwargs, **kwargs}
+    return _process_strategy(
+        data,
+        internal_cols=quadruple_strike_internal_cols,
+        external_cols=quadruple_strike_external_cols,
+        leg_def=leg_def,
+        rules=_rule_iron_condor_strikes,
+        join_on=["underlying_symbol", "expiration", "dte_entry", "dte_range"],
+        params=params,
+    )
+
+
+def _iron_butterfly(
+    data: pd.DataFrame, leg_def: List[Tuple], **kwargs: Unpack[StrategyParams]
+) -> pd.DataFrame:
+    """Process iron butterfly strategies (4 legs, middle legs share strike)."""
+    params = {**default_kwargs, **kwargs}
+    return _process_strategy(
+        data,
+        internal_cols=quadruple_strike_internal_cols,
+        external_cols=quadruple_strike_external_cols,
+        leg_def=leg_def,
+        rules=_rule_iron_butterfly_strikes,
+        join_on=["underlying_symbol", "expiration", "dte_entry", "dte_range"],
+        params=params,
+    )
+
+
+def _covered_call(
+    data: pd.DataFrame, leg_def: List[Tuple], **kwargs: Unpack[StrategyParams]
+) -> pd.DataFrame:
+    """
+    Process covered call strategy.
+
+    Note: This implementation simulates a covered call using options data only,
+    approximating the underlying position through the relationship between
+    option premiums and underlying price changes.
+    """
+    params = {**default_kwargs, **kwargs}
+    return _process_strategy(
+        data,
+        internal_cols=double_strike_internal_cols,
+        external_cols=double_strike_external_cols,
+        leg_def=leg_def,
+        rules=_rule_non_overlapping_strike,
+        join_on=["underlying_symbol", "expiration", "dte_entry", "dte_range"],
+        params=params,
+    )
+
+
+def _calendar_spread(
+    data: pd.DataFrame,
+    leg_def: List[Tuple],
+    same_strike: bool = True,
+    **kwargs: Unpack[StrategyParams],
+) -> pd.DataFrame:
+    """
+    Process calendar or diagonal spread strategies.
+
+    Calendar spreads have the same strike but different expirations.
+    Diagonal spreads have different strikes and different expirations.
+
+    Args:
+        data: DataFrame containing option chain data
+        leg_def: List of tuples defining strategy legs [(side, option_filter), ...]
+        same_strike: True for calendar spreads, False for diagonal spreads
+        **kwargs: Optional strategy parameters
+
+    Returns:
+        DataFrame with calendar/diagonal spread strategy results
+    """
+    params = {**calendar_default_kwargs, **kwargs}
+    internal_cols = (
+        calendar_spread_internal_cols if same_strike else diagonal_spread_internal_cols
+    )
+    external_cols = (
+        calendar_spread_external_cols if same_strike else diagonal_spread_external_cols
+    )
+
+    return _process_calendar_strategy(
+        data,
+        internal_cols=internal_cols,
+        external_cols=external_cols,
+        leg_def=leg_def,
+        rules=_rule_expiration_ordering,
+        same_strike=same_strike,
+        params=params,
+    )
